refactor(quiz): split filter UI + config hook to satisfy size limits
Adding the image-filter toggles pushed question-filters.tsx (181) and
use-quiz-config.ts (94) over the 150/80-line limits. Extract the filter
primitives + toggle constants into question-filter-toggles.tsx, and the
selection state/handlers + count-sync effect into use-quiz-config-state
and use-filtered-count-sync. Behavior unchanged. Addresses code-reviewer
BLOCKING findings on #864.
